Move OmniSharp activation out of main
